Optimised not_mersenne algorithm
This optimisation is assuming that digit.bit_length() is a
constant-time optimisation; thanks to __builtin_clz, it should
be in GCC/C18.

This works by counting how long it'll take to get up to digit,
bit-shifting up to there, then doubling again if we're not
close enough yet. The bit shift will definitely get almost
there, which I could prove but am too tired to at the moment;
that's why it just needs an if instead of a while.

Then, since we're between base and 2*base, subtracting base is
a faster way of doing modulo. Then, and only then, have we got
a chance of digit being 1, since that's the only place in which
there's any subtraction,* so it's fine to only check there.

* assuming that at the start of the loop, digit is always less
than base; I haven't proven this but it IS the case!