Replaced primes with a sieve.
This code is the product of a tired, tired mind. For some
reason, I decided not to add ANY sane comments to this. So,
here's my explanation.

The first check: output_index is set to -2 because when it's
increased by 1 it needs to say -1 so that it compares bigger
than calculated. Don't ask.

As for why calculated is 2 instead of 1? Well, I needed to have
the biggest known prime be 3, because the code relies on all
primes being odd but never uses the 2 value anywhere so we're
good.

Now, this code will eventually stack overflow. I don't know why
I wrote it that way. There's something seriously wrong going
on here. But with a bit of luck I'll be able to multiply the
number of numbers it can calculate until it does by sqrt(2) or
thereabouts using the cool (i + i>>1) trick I thought of.
That's *1.5, if you must know. Yes, yes, premature optimisation
but I really need it to round down which *3/2 would do anyway,
but… Ok, I'll check whether the compiler does anything clever
with (*3)/2 before I implement my cool hack.

Where was I? Oh yes. The lower number needs to be 2 higher than
the biggest prime, but the upper number needs to be the square
of the biggest prime. The solution? Add 2 after using the value
to square… That can't be fast in a pipelined processor. I
really, really need to start trusting the compiler more.

Anyway. sieve_arr is an array, and I do pointer tricksiness to
avoid having to subtract every time I index the array by
getting a pointer to the projected 0 index of the conceptual
array (of which I've only allocated a portion) instead.

Then, instead of using the standard library, I write my own…
ha haha hahaha! I write my own loop, with a comment of // zero,
that sets the values to 1! This is PRICELESS!

Ok, I've calmed down now. Now for the actual sieve. You'll see
that I've used a ternary here. That's a last-minute kludge to
fix an error that caused it to skip over square numbers if they
were the last excluded value. Basically, at the… Oh. Well, I
could've just explicitly set the first one to false earlier,
in the // zero section, seeing as it's certainly a square…

That's another thing on my todo list.

Anyway (which has apparently become my favourite word!), this
is just a simple sieve. The core code.

Now, I'm not proud of the next bit, but I see no other way of
doing it. I loop through every item, counting them all up, and
then realloc that much memory. (I think there's no off-by-one
error here, but I'm not certain. Todo todo todo.)

When I said that I had no sane comments, I lied. I have one.
The // save section saves the numbers to the oh wow that's bad.
Sorry. But this code is unreadable. I mean, it works, but there
are so many things wrong with it.

Ahem.

The reason for the *(++destination) part is that
&known_primes[output_index] is the location of the current
highest known prime number, and we don't want to overwrite it.

Concluding message: writing good C code is still hard.